// src/models/regions/locations/spawn/blueprints/buildings/types.ts
import { Point } from '../../../../../utilities/math/points'

type workshops =
  | 'cobbler'
  | 'furniture maker'
  | 'furrier'
  | 'weaver'
  | 'basket weaver'
  | 'carpenter'
  | 'paper maker'
  | 'potter'
  | 'wheelwright'
  | 'jeweler'
  | 'mason'
  | 'baker'
  | 'soap maker'
  | 'chandler'
  | 'cooper'
  | 'pastry maker'
  | 'scabbard maker'
  | 'silversmith'
  | 'saddler'
  | 'purse maker'
  | 'blacksmith'
  | 'goldsmith'
  | 'toy-maker'
  | 'artist'
  | 'leatherworker'
  | 'rope maker'
  | 'tanner'
  | 'buckle maker'
  | 'cutler'
  | 'fuller'
  | 'harness maker'
  | 'painter'
  | 'woodcarver'
  | 'glass maker'
  | 'instrument maker'
  | 'locksmith'
  | 'rug maker'
  | 'sculptor'
  | 'bleacher'
  | 'shipwright'
  | 'bookbinder'
  | 'bowyer/fletcher'
  | 'brewer'
  | 'glove maker'
  | 'vintner'
  | 'girdler'
  | 'skinner'
  | 'armorer'
  | 'weaponsmith'
  | 'distiller'
  | 'illuminator'
  | 'perfumer'
  | 'tiler'
  | 'clock maker'
  | 'vestment maker'
  | 'alchemist'
  | 'bell maker'
  | 'dye maker'
  | 'inventor'

type shops =
  | 'clothier'
  | 'grocer'
  | 'dairy seller'
  | 'launderer'
  | 'furrier'
  | 'tailor'
  | 'barber'
  | 'draper'
  | 'flower seller'
  | 'jeweler'
  | 'mercer'
  | 'engraver'
  | 'pawnbroker'
  | 'haberdasher'
  | 'wine merchant'
  | 'tinker'
  | 'butcher'
  | 'fishmonger'
  | 'wool merchant'
  | 'beer merchant'
  | 'herbalist'
  | 'spice merchant'
  | 'wood seller'
  | 'brothel keeper'
  | 'hay merchant'
  | 'bookseller'
  | 'religious souvenir seller'
  | 'naval outfitter'
  | 'grain merchant'

type offices =
  | 'livestock merchant'
  | 'carpenter'
  | 'mason'
  | 'pawnbroker'
  | 'wine merchant'
  | 'surgeon, unlicensed'
  | 'wool merchant'
  | 'beer merchant'
  | 'spice merchant'
  | 'surgeon, licensed'
  | 'moneychanger'
  | 'scholar'
  | 'advocate'
  | 'historian'
  | 'engineer'
  | 'architect'
  | 'astrologer'
  | 'grain merchant'
  | 'banker'
  | 'cartographer'

export interface Building extends Point {
  idx: number
  type:
    | "lord's manor"
    | "lord's keep"
    | 'royal palace'
    | 'residence'
    | 'government'
    | 'asylum'
    | 'barrack'
    | 'bath'
    | 'lodging'
    | `cemetary (${'small' | 'large'})`
    | 'religious'
    | 'cistern'
    | 'coliseum'
    | 'docks'
    | 'gates'
    | 'livestock'
    | 'fountain'
    | 'garden'
    | `park (${'small' | 'large'})`
    | 'granary'
    | 'guild house'
    | 'library'
    | 'mill'
    | `office (${offices})`
    | 'plaza'
    | 'prison'
    | `shop (${shops})`
    | 'stable'
    | 'tavern'
    | 'tenement'
    | 'theater'
    | 'university'
    | 'warehouse'
    | 'well'
    | `workshop (${workshops})`
  quality: { grade: 'A' | 'B' | 'C' | 'D'; desc: string }
}